Refactor trial session eligible cases integration tests into one
